Add mobile_client.get_top_songs() to "Liked Songs"
The current list of liked songs is only populated with those that happen
to have been enumerated, leaving out songs that have been marked as
thumbs up but are not in the library or a playlist.  This patch
explicitly walks the list from "get_top_songs()", which contains some of
the songs that have been marked as thumbs up.  The result is that my
liked songs list matches what is shown on my phone.

See https://github.com/simon-weber/gmusicapi/issues/637 for more
information.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>